% Auto-generated: do not edit by hand
\name{ethaddress}

\alias{ethaddress}

\title{ethaddress component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
ethaddress(id=NULL, address=NULL, textLabels=NULL, key=NULL)
}

\arguments{
\item{id}{Character. The ID of this component, used to identify dash components
in callbacks. The ID needs to be unique across all of the
components in an app.}

\item{address}{Character. Eth Address.}

\item{textLabels}{Logical. The children of this component}

\item{key}{Character. A unique identifier for the component, used to improve
performance by React.js while rendering components
See https://reactjs.org/docs/lists-and-keys.html for more info}
}
