% Auto-generated: do not edit by hand
\name{metamask}

\alias{metamask}

\title{metamask component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
metamask(id=NULL, value=NULL, borderRadius=NULL, boxShadow=NULL, width=NULL,
className=NULL, key=NULL)
}

\arguments{
\item{id}{Character. The ID of this component, used to identify dash components
in callbacks. The ID needs to be unique across all of the
components in an app.}

\item{value}{Character. }

\item{borderRadius}{Numeric. Defines CSS styles which will override styles previously set.}

\item{boxShadow}{Numeric. The children of this component}

\item{width}{Numeric. Often used with CSS to style elements with common properties.}

\item{className}{Character. Often used with CSS to style elements with common properties.}

\item{key}{Character. A unique identifier for the component, used to improve
performance by React.js while rendering components
See https://reactjs.org/docs/lists-and-keys.html for more info}
}
